mcp-server/requirements.txt
# MCP Server & Todo Tooling - Python Dependencies
# Python 3.11+

# MCP Protocol
mcp>=1.0.0

# Database
sqlmodel>=0.0.14
asyncpg>=0.29.0

# Authentication
pyjwt>=2.8.0
bcrypt>=4.1.2

# Configuration
python-dotenv>=1.0.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Development
black>=23.12.1
ruff>=0.1.9
